+++ tests/testthat/test-parsnip.R
@@ -90,7 +90,8 @@ test_that("ggemmeans, parsnip", {
 })
 
 test_that("test_predictions, parsnip", {
+  skip_on_os("linux")
   out <- test_predictions(fit, "c172code")
   expect_equal(out$Slope, 0.71836, tolerance = 1e-3)
-  expect_equal(out$conf.low, -1.92486, tolerance = 1e-3)
+  expect_equal(out$conf.low, -1.928975, tolerance = 1e-3)
 })