+++ NEWS.md
@@ -22,6 +22,10 @@
 
 * `hypothesis_test()` nwow suppports models from package *parsnip*.
 
+* `johnson_neyman()` gains a `p_adjust` argument, to adjust p-values for multiple
+  comparisons. Currently, only `p_adjust = "esarey"` (resp. `p_adjust = "es"`) and
+  `p_adjust = "fdr"` (resp. `p_adjust = "bh"`) are supported.
+
 ## Bug fixes
 
 * Fixed issue in `ggpredict()` for `type = "random"` when sampling from random